requirements-dev.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.11.5
configobj==5.0.6
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
decorator==4.1.2
dj-database-url==0.5.0
dj-static==0.0.6
enum34==1.1.6
#functools32==3.2.3.post2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
ipython==5.5.0
ipython-genutils==0.2.0
lxml==4.2.3
#nemo-emblems==3.8.0
parsel==1.5.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.2.0
#pygobject==3.26.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
#pyparted==3.11.1
#python-apt==1.6.2
python-decouple==3.1
python-xlib==0.20
pyxdg==0.25
queuelib==1.5.0
scandir==1.7
#Scrapy==1.5.1
#selenium==3.14.0
service-identity==17.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
traitlets==4.3.2
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
wcwidth==0.1.7
zope.interface==4.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
Pillow==5.1.0
python-decouple==3.1
pytz==2018.3
static3==0.7.0





